feat: project routes are organized.
